Add a method to get the closest restaurants

+++ src/main/java/org/masonord/delivery/dto/model/GeoCodingDto.java
@@ -13,7 +13,7 @@ import lombok.experimental.Accessors;
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class GeoCodingDto {
-    float lat;
+    double lat;
 
-    float lon;
+    double lon;
 }
